test_document_ai/requirements.txt
google-cloud-documentai
google-auth
google-api-core